backend/requirements.txt
# ===============================
# FastAPI Core
# ===============================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ===============================
# News Collection / Scraping
# ===============================
feedparser==6.0.10
requests==2.31.0
urllib3==2.1.0
beautifulsoup4==4.12.3

# ===============================
# AI / ML (Google)
# ===============================
google-genai
google-generativeai
google-cloud-texttospeech
google-cloud-speech==2.24.1

# ===============================
# Core Utilities
# ===============================
python-dotenv==1.0.0
python-dateutil==2.8.2

# ===============================
# Database
# ===============================
psycopg2-binary==2.9.9

# ===============================
# Scheduler
# ===============================
schedule==1.2.1
apscheduler==3.10.4
croniter==1.4.1

# ===============================
# AWS S3
# ===============================
boto3==1.34.0

# ===============================
# Media & Image Processing
# ===============================
Pillow==9.5.0
moviepy==1.0.3
ffmpeg-python==0.2.0

# ===============================
# Social Media & Automation
# ===============================
telethon==1.34.0
playwright==1.41.2

# ===============================
# Arabic Text Support (RTL)
# ===============================
arabic-reshaper==3.0.0
python-bidi==0.4.2
